Move map color scale into seperate component
