+++ client/src/components/sharedComponents/map/colorScaleColors.ts
@@ -1,0 +1,30 @@
+const startColor = '#dc4900';
+const endColor = '#145500';
+
+export const twoColorScale: [string, string] = [
+  startColor,
+  endColor,
+];
+export const fiveColorScale: [string, string, string, string, string] = [
+  startColor,
+  '#cb9e00',
+  '#99b900',
+  '#4a8900',
+  endColor,
+];
+export const thirteenColorScale:
+  [string, string, string, string, string, string, string, string, string, string, string, string, string] = [
+  startColor,
+  '#d37700',
+  '#ce9200',
+  '#cb9e00',
+  '#c8aa00',
+  '#c5b500',
+  '#b5bf00',
+  '#99b900',
+  '#8ab100',
+  '#7ca900',
+  '#629900',
+  '#4a8900',
+  endColor,
+];
